04/05/09 작업사항 commit

+++ domain/src/main/java/hee/study/domain/usecase/DeleteTodoUseCase.kt
@@ -4,5 +4,5 @@ import hee.study.domain.repository.TodoRepository
 import javax.inject.Inject
 
 class DeleteTodoUseCase @Inject constructor(private val repository: TodoRepository) {
-    suspend fun execute(id: Int) = repository.deleteTodo(id)
+    suspend fun execute(id: Long) = repository.deleteTodo(id)
 }
